refactor: EditHeader 컴포넌트로 이름 변경
