+++ gitlio/app/(edit)/experience/page.tsx
@@ -1,5 +1,5 @@
 import EditBuilder from "@/components/EditBuilder";
-import Header from "@/components/Header";
+import EditHeader from "@/components/EditHeader";
 import SideBar from "@/components/SideBar";
 import Title from "@/components/Title";
 
@@ -10,7 +10,7 @@ export default function ExperiencePage() {
         <div className="flex justify-center col-span-1 row-span-9 h-full border border-blue-400">
           <SideBar />
         </div>
-        <Header />
+        <EditHeader />
         <div className="col-span-4 row-span-8 h-full bg-[#DCE6FF] border border-blue-400">
           <div className="flex flex-col h-full items-center px-5 ">
             <Title />
