+++ gitlio/components/EditHeader.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+import PublishBtn from "./PublishBtn";
+
+export default function EditHeader() {
+  return (
+    <div className="flex items-center justify-end col-span-4 row-span-1 w-full h-full bg-[#DCE6FF] border border-blue-400 px-4">
+      <PublishBtn />
+      <div className="ml-5 w-16 h-16">
+        <img className="w-full h-full rounded-full" src="/gdsc.jpg"></img>
+      </div>
+    </div>
+  );
+}
